Migrate float -> Decimal

+++ src/trading_bot/models/kline_event.py
@@ -1,14 +1,16 @@
 from dataclasses import dataclass
 from datetime import datetime
+from decimal import Decimal
+
 
 #TODO float -> Decimal
 @dataclass
 class KlineEvent:
     event_time: datetime | None
     open_time: datetime
-    open: float
-    high: float
-    low: float
-    close: float
-    volume: float
+    open: Decimal
+    high: Decimal
+    low: Decimal
+    close: Decimal
+    volume: Decimal
     is_closed: bool
